move WrapMode into core/geometry/curve.ts
