+++ cocos/core/geometry/index.ts
@@ -39,7 +39,7 @@ export { AABB } from './aabb';
 export { OBB } from './obb';
 export { Capsule } from './capsule';
 export { Frustum } from './frustum';
-export { Keyframe, AnimationCurve } from './curve';
+export { Keyframe, AnimationCurve, WrapModeMask } from './curve';
 export { SplineMode, Spline } from './spline';
 export * from './spec';
 export * from './deprecated-3.0.0';
